dev: refactor root scripts (#714)
Converts most `scripts/` from JS to TS.

The script for updating deps has been removed and we just use `taze`
now, instead of a wrapped npm-check-updates. `taze` has an interactive
mode and supports pnpm catalogs out of the box.

Also the consistent-package-json script no longer tries to write to the
pnpm-workspace.yaml catalogs. Instead it will only update the necessary
package.json files if a catalog-able dep is found. This also undoes some
of the cataloged deps that were added by the old behavior.

There is a new `find-catalogable-deps` script that will report deps that
could be cataloged, but there are too many heuristics for this to be an
automatic process. So this script can be used to find some and then
manually catalog them.