release: notify slack from workflow for important alerts (#721)
The release PR should always be green because we test our PRs before
merging which should also be green. And then merging the release PR
should always successfully publish the packages.

But these "always" things can fail for some of the following reasons.
This adds slack notifications for these cases to a new slack channel
`#github-alerts`. A slack app and webhook URL has already been created
and added to this repo as a secret.

### Flaky CI
We've mostly squashed these for now but we want to be alerted if they
start to happen again. Currently if these do happen they are easy to
ignore because the PR will be recreated pretty often.

### Not Rebasing PRs Before Merge
We tried turning on a rule for this on PRs but it slowed us down a lot.
I think this is still the right choice but because of the nature of a
monorepo with many workspaces, not rebasing can cause failures on main
after merging. Here's a recent commit where that happened:
c133f87ee96f915d01170db8815a36c3f5e43e92

### Release PR Fails to Create/Update
This error would be mostly invisible until someone went to check the
release PR and noticed it was behind main.

### Publish Fails After a Merged Release
A release can also fail to be published after a successful release PR is
merged. In that case it is very important to follow up manually to see
what went wrong.