gui: badge and icon updates (#731)
### `<DataBadge />` and `<InlineCode />`
This PR creates a new `data-badge` component to use in places where a
badge with data is required, i.e.: in metadata or smaller badges with
counts, and allows us to remove the `copy-to-clipboard` component which
was used in only a couple of places.
- It differs from `inline-code`, which should be used in places where
literal 'inline code' should be used.
- It simplifies `inline-code` and removes the `tooltip` api from it.
- It removes the need for `copy-to-clipboard`, as it has this
functionality built-in

### Icons
This PR updates some icons that relied on svgs, to be constructed from
the `createLucideIcon` utility function from `lucide-react`.
- It makes styling these icons with lucide props easier and more
consistent
- It allows us to remove the extra nerdFont `zed-mono` dependency which
was used for `glyph-icon`
- It removes the `glyph-icon` component

### Updates
- Updates places where `<InlineCode />` is to be replaced with
`<DataBadge />`
- Updates places where the new icon styling is required
- Removes `zed mono` font from `index.html` and declarations from
`main.css`

### Bugfixes
- It corrects a typo in `index.html` where the `ref` attributes on
`links` should've actually been `rel`