query: fix scope selector (#734)
The `:scope` selector should default to `:root` but its value should be
parameterized, so that we can potentially set it to a given workspace
when running from a workspace folder.

Also, `vlt query` and `vlt ls` should set their default scope to either
the root folder or a workspace folder dependending on the context.

Fixes: https://github.com/vltpkg/vltpkg/issues/718