cli: improve how invalid options are reported
This applies the same `code: 'ECONFIG'` behavior for all config options,
including invalid values, providing more assistance beyond just listing
the available options.

Now it properly detects for example when an option is given a value that
shouldn't be (because it's a flag) or vice versa (not given a value to
an option that is not a boolean flag), or given an invalid value (eg,
`--config=asdfasdf` where only specific values are allowed).

It also properly differentiates between config _file_ errors and cli
option errors, presenting the suggestions in a more useful manner.