spec: default npm: alias to configured default registry
If you're setting your default registry to a caching proxy, it'll be
very surprising to have any public-registry packages using the npm:
alias to be fetched from the public registry, which your system might
not even be able to access.

This fixes the problem by making the npm: alias default to the
*configured* default registry (which of course defaults to the default
public registry), so there's no change in the normal case, but it'll
stop being broken in the cause where the public registry is pointed
somewhere else, without requiring the user to *also* set the npm: alias
when changing the default registry.

Found while writing documentation about working with registries.

Note that this has to be done a bit carefully, or else we'll think that
a package is referencing the `npm:` alias when it's a transitive
dependency of a dep coming from a named alias registry.