ci: exclude node 24 for now [ci full-matrix] (#759)
GUI tests fail on Node 24. This is not an issue for the actual GUI
but with the tooling required to run the tests in Node (vitest, etc).
Found a similar report here:
https://github.com/reduxjs/redux-toolkit/issues/4966
For now just exclude it from the matrix always, we'll add it back later
once tooling catches up.