query: fix names in attribute selector
After the previous id selector refactor it has been stablished that
in order to provide a more concise and unambiguous experience the
id selectors are the only way to select by edge name and the attribute
selector should now only match in properties from the node manifest.

Based on that this changeset removes the affordance to match edge
names in the attribute selector.

Refs: https://github.com/vltpkg/vltpkg/pull/643