fix: Handle jsr: without a name like yarn, pnpm do
That is, `@a/b@jsr:1.x` should be the same as `@a/b@jsr:@a/b@1.x`.

This also updates the save behavior, so that if you install
`@a/b@jsr:@a/b@1`, it'll save as `"@a/b": "jsr:1"`, which is a lot
cleaner and more in line with the other implementations.