gui: version count in `tabs-versions` (#772)
https://github.com/user-attachments/assets/499baf80-a42b-44fa-9eca-666e50903ea7

Closes vltpkg/statusboard#152

This PR adds a new `number-flow` component to the gui, which is used for
animating between a 'start' and 'end' value, and integrates it into the
`@/selected-item/tabs-versions`'s `VersionHeader` component to nicely
animate the amount of versions that have been filtered when a user
applies a filter or searches in the search input.

This PR also adds some perfomance improvements by memoizing filter
functions and components inside the `tabs-versions` which rely on the
`versions` prop to optimize re-renders. It also removes some layout
animations from motion to improve the stability of the animations.