gui: separate grid results and selected item components (#775)
There should be no behavior or visual changes in this PR. The
`<ExplorerGrid>` is now a smaller wrapper which will get the list of
items from the query and either render a `<SelectedItem>` or a
`<Results>`. Both of these new components require the data to be passed
now via props instead of getting it from the store.

This is the basis to then make changes that will only be in `<Results>`.
Ref https://github.com/vltpkg/statusboard/issues/111