chore: update to pnpm@10
This is so we can use the `jsr:` prefix instead of an `.npmrc` file
which will make it easier to self install with `vlt` later.

This also adds a `packageManager` field to the root package.json so
Vercel will use the correct version of `pnpm`. AFAICT this is the only
way to achieve this.