graph: fix edge replacement
Replacing existing edges was resulting in missing nodes in the graph
since the `removeNodes` method removes any edges going into the node
that is about to be removed.

A straightforward and resilient solution is to make sure we remove
any edge that is about to have its destination replaced from the
`edgesIn` set in that original destination node.

Refs: https://github.com/vltpkg/statusboard/issues/64