gui: dependency item design + filtering (#787)
https://github.com/user-attachments/assets/72e8cbe6-8c03-48c2-912e-51b9a3388047

### Overview
This PR updates the `side-item` design in the gui by 'compacting' them
from a 2-tier card to a single card. It introduces a context menu around
the `side-item` to replace the ellipses.

Additionally it simplifies the logic for rendering the 'connector' lines
for the side items by relying on pure tailwind classes instead of
`useRefs` and `useEffects` since the cards will only be a single tier
now.

It creates a filtering system on to the `dependency-sidebar` and creates
`filter/` components to allow the user to filter through the
`dependencies` based on their 'relationship' (i.e.: prod, dev ...).
It integrates the `number-flow` component into the `dependency-sidebar`
headers to visualize the number of dependencies that were filtered.