gui bugfix: versions tab filtering (#790)
Closes vltpkg/statusboard#166

### Bug Overview

When applying a filter or searching through the `versions` tab list, the
versions are correctly filtered and rendered in order. However when a
user scrolls through the list (in such a way to load more versions), the
order of filtering for the list is lost and the versions are no longer
in the correct order.

This was due to the logic with the `intersectionObserver` which was not
correctly handling the order of the versions when new ones were loaded.

### Bugfix

In order to create a more robust solution to handle the perfomance
issues with long lists, the `intersectionObserver` logic was removed and
replaced with (`react-virtuoso`)[https://virtuoso.dev/] which allows us
to virtualize the list of versions, improving the performance and
ensures that the order of the versions is maintained when scrolling.

### Smaller Changes

Creates a 'blurred' edge around the top and bottom of the `versions`
tab's version list for aesthetics.