dss-breadcrumb: interactive refactor (#816)
Refactored the breadcrumb module to better manage state and the parsed
result.

The exported class that represents the parsed result of a query is now
named `Breadcrumb`, instances of this class now have a `interactive()`
method that returns an `InteractiveBreadcrumb` which is now the object
that is going to hold state on the current item and implements a
`next()` method that allows for updating that state.

Refs: https://github.com/vltpkg/statusboard/issues/64