gui: bugfix tooltip content overlay (#827)
| Before | After |
| --- | --- |
| <img width="100%" alt="Screenshot 2025-06-04 at 10 59 18"
src="https://github.com/user-attachments/assets/23204ddb-8c46-4faf-8c63-eb1c03b097b6"
/> | <img width="100%" alt="Screenshot 2025-06-04 at 12 54 19"
src="https://github.com/user-attachments/assets/0dc42eec-484e-40bc-944f-a11bee4c21ba"
/> |

This PR fixes the issue with the `sidebar`'s link tooltips incorrectly
being rendered on the `explorer` view.

### Bug Explanation
When we updated the `query-bar` component to have an absolute position
in order to stick to the top of the screen when scrolling, the z-indexes
collided with tooltips in the gui which were rendered on the `explorer`
view.

### Bug Fix
Exported a primitive `TooltipPortal` from `Radix` in `tooltip.tsx` and
wrapped the `TooltipContent` with it. This allows the content to be
rendered on the `document.body` instead of being constrained by the
parent element's z-index.

### Other Changes
- updates styles on `explorer` view
- adds a `tooltip` to the `filter` button