graph: add graph modifier try dependencies helper
Add a new `GraphModifier.tryDependencies` helper that takes a few of
the repetitive boilerplate from the graph traversal into the modifiers
helper module.

This also changes the `GraphModifier.tryNewDependency` method so that
it only returns a single matching active modifier entry, which entry
to return is defined by first selecting from any breadcrumb that is
already complete and then using specificity to define which query
should be prioritized.

Refs: https://github.com/vltpkg/statusboard/issues/64