dss-breadcrumb: add support to  `:semver` & `:v` selectors  (#881)
`ModifierBreadcrumbItem` objects will now expose a `comparator()` method
that will allow for checking extra values, enabling the usage of a more
diverse set of pseudo selectors.

This implements the initial `:semver` and its shortcut `:v` to allow for
fine tuning breadcrumb matching based on the versions of packages and
user defined semver values, e.g:

    :root > #eslint > #minimatch:v(3)

Also refactors parsing so that any items before / after a combinator
are treated as an unique item and not only the specific cases
of the `:workspace` & `:project` pseudo selectors.

Refs: https://github.com/vltpkg/statusboard/issues/176