cli-sdk: add --scope to pkg, ls, query
Add support to the `--scope` option to the following commands:
- `vlt pkg`
- `vlt ls`
- `vlt query`

In `vlt pkg` the `scope` option will now allow for retrieving, setting
and removing keys from the `package.json` file of any packages found
using the provided `scope` query.

In `vlt ls` and `vlt query` using the `scope` option allows for
setting arbitrary roots for selecting a sub-graph that can then be
visualized in all the supported formats.