server: load data from hidden lockfile
The GUI server now always load data from the hidden lockfile instead of
loading from traversing the `node_modules` folder. This ensures it has
access to the `ManifestRegistry` object that is returned from the
registry which contains additional data such as `dist` and a complete
set of `contributors` which is normalized from the `maintainers` data.

Refs: https://github.com/vltpkg/vltpkg/pull/977
Refs: https://github.com/vltpkg/vltpkg/pull/947