dep-id: dry the dep-id module (#998)
At some point we duplicated all the implementation in the
`@vltpkg/dep-id` module to provide an isomorphic `/browser` variation
that does not rely on node internals.

This has been a liability since the node-specific implementation kept
evolving without updating the isomorphic implementation in
`./browser.ts`.

This refactors it to make sure there's proper reusability while only
redefining the specific methods that rely on the node internals:
`hydrate` and `hydrateTuple`.