cli-sdk(pack, publish): support publish directory (#1002)
This PR adds a `--publish-directory` config which when set will use that
directory as the cwd for pack and publish. It is designed so that
lifecycle scripts like `prepare` can populate that directory with files
from a bundler or compiler.

With this implementation it is possible to do everything that was
possible with `publishConfig` and since we use pnpm's
`publishConfig.directory` to do the same thing currently we only need
the directory feature for dogfooding. I think this makes more sense as a
base feature that we build on top of in the future, without needing to
support `publishConfig` (yet?).