gui: smoother tab transitions (#1000)
https://github.com/user-attachments/assets/83f16dc7-a1c6-487c-a967-3ca2cb51c285

### Overview

This PR updates the `tab` based navigation in the `selected-item` with
smoother transitions. Notably it removes the `tab` state's from
`selected-item` `context` and treats the `URL` as the source of truth
instead. It creates a new `useTabNavigation` to handle the tab
navigation logic, and creates `helpers.tsx` inside `selected-item/`
directory to supercharge the `tab` components from `shadcn`.


#### Changelog
- **feat: removes stateful tab implementation**
- **feat: implements smoother transitions with `motion`**
- **tests: updates tests and snapshots**