normalize manifest keywords (#1039)
## Overview

This PR normalizes the `keywords` in the manifest.

Previously some keywords could've been strings separated by commas or
contain some invalid keywords that weren't being validated, such as `"
"` or `[" "]` and `[""]`.

This is partially linked to the issue in #1036

Additionally this PR unlocks being able to query by the normalized
keywords attribute with predictable results