build: remove compiled binary from smoke testing and publishing (#1059)
This leaves most of the functionality intact that will be needed to
compile
a binary and publish it in the future, since it contains some logic that
will be reused when that time comes.

But we don't know how that binary will be compiled so this PR also
removes
`deno` setup from our CI process.

And it removes running the compiled binary from the smoke test as well
since that was not testing anything that ended up in front of users.