server: create `/fs/ls` endpoint (#1111)
### Overview

This PR creates a new endpoint in the `server` workspace, `/fs/ls` to
allow interfacing with the filesystem. This endpoint will be used to
list files and directories in a specified path.

This is also a pre-requisite for #1025 and vltpkg/statusboard#71, which
will implement a ui to view files and directories, allowing a user to
choose a default `dashboard-root` path.

#### Example Request `/fs/ls`

```json
{
  "path": "/path/to/directory"
}
```