fix(spec): correctly load github PRs
Tarballs from GitHub PRs can only be loaded from the /tarball endpoint
from the GitHub API, not using the codeload.github.com URLs.

So this PR changes the host archive URL for `github` to use the API
which allows for all other types of `committish` loading as well.

There is one caveat, which is that in order to allow the format of
`org/repo#pull/number` we have to append `/head` at the end of the URL
as long as there is not already another `/path` at the end since the
user might want to install the merge commit with `#pull/number/merge`.