docs: fix icons (#1167)
## Overview

There were bugs with the icons on the sidebar, where clicking to expand
a group would rotate all icons in that group (including the main icon)

Additionally the logic for the icons was pretty verboise and repetitive,
this PR makes them drier by simplifying the sidebar menu and moving the
majority of the list info into `@/components/sidebar/data` to keep this
logic separated from the ui components.

It prefers to use icons created through `lucide-reacts`
`createLucideIcon` method, as it optimizes the compatibility and makes
styling easier across all of our platforms. Icons in `public/icons` were
removed and new icons were created in `@/components/icons/` with the
`createLucideIcon` method.

### Before

https://github.com/user-attachments/assets/00141879-c575-4149-8da6-232eb1652c9d

### After

https://github.com/user-attachments/assets/1bd0488a-ed26-42eb-b3e5-5ad8899791d0

### Changelog
- **feat: prefers to use `createLucideIcon` instead of `<svg>`**
- **feat: simplifies sidebar menu**
- **chore: fixes references to old `svg` icons**
- **chore: lint and format**

---------

Co-authored-by: Luke Karrys <luke@lukekarrys.com>