gui: introduce a `/settings` view (#1025)
## Overview
This PR introduces a new `/settings` route and view in the GUI.

https://github.com/user-attachments/assets/cc28b538-0de9-4edc-8b19-b28b0994b5db

It integrates the `file-explorer` components to allow a user to select a
`dashboard-root` from the `/settings` route by creating new interfaces
in `@/lib/` (`vlt-config.ts` and `fetch-fs.ts`) that interact with the
`/fs` and `/config` endpoints in the server workspace.

It also introduces a new `alert-dialog` component that displays on the
`/dashboard` route when there is no `dashboard-root` configured (check
done by `use-dashboard-root-check` hook in `@/components/hooks`)
prompting the user to either use the default homedir or select a
directory handle on their machine.

When there is no projects to be displayed in the configured
`dashboard-root` it presents an empty state prompting the user to create
a new project in that directory, or update their `dashboard-root`
configuration.

It also moves the `theme-switcher` component into the `/settings` view
to free up some space on the sidebar.

### Generic Breadcrumb Header
This PR consolidates the `/help` header into a new `breadcrumb-header`
component used in the `header` tree, and applies the same header to the
`/settings` route.

### Sidebar Update
The sidebar gets a new contextual `setting` component that is only
rendered in the `/settings` route, and provides quick navigation between
settings.

### Setting Components
This PR introduces new components related to settings, and provides some
useful default inputs and composables for the settings view.

### File Explorer Components
This PR integrates the `file-explorer` components for the
`dashboard-root` selection