query: add `:host-context` selector (#1197)
Add a new `:host-context()` selector to the DSS query language.

This new selector will allow changing the context of the query to target
a new context.

It's possible to target different local folders as valid contexts using
the following syntax:

```
:host-context("file:~/tmp/my-project")
```

A special context that refers to all projects that can be found in the
configured user dashboard named `local` is also available:

```
:host-context(local)
```