gui: results view sorting and pagination (#1198)
## Overview

This PR updates the results view in the gui to allow for pagination of
results and sorting abilities.

<img width="1852" height="1087" alt="Screenshot 2025-09-12 at 15 14 30"
src="https://github.com/user-attachments/assets/6543e1f9-99bc-4419-8466-fd92ff73e2e2"
/>

It allows the sorting, per page size of results, and current page to be
stored as URL params in the form of:
`/explore/<compressedQuery>?page=1&pageSize=25&sort=sortingOption` for
easy sharing and backlinking.

It creates a new scoped zustand store in context for the `results`, and
allows the results components to share that state (similar to how we
implemented the `selected-item` zustand context)

### Other Changes
- Rearranges some imports in `routes.tsx` for consistency
- Adds some code documentation to the `use-query-navigation` hook
- Updates the `tabs` styling once again in `selected-item` due to a line
overflow bug

### Changelog
- **chore: adds some code documentation**
- **chore: rearranges some imports**
- **chore: updates the tabs styles**
- **feat: adds `pagination` component from `shadcn`**
- **feat: updates `results` view**
- **feat: creates a scoped zustand context store for `results`**
- **feat: adds `page-navigation` component for `results`**
- **feat: adds `page-options` components for `results`**
- **feat: adds `sort` component for `results`**