cli-sdk: add vlt build command
This changeset fundamentally changes how install works in the
vlt cli by breaking installs into two distinct steps:
- Install step that resolves a graph data structure from
parsing `package.json` files retrieved and extracts tarball
data into the `node_modules` folder.
- A build step that runs the lifecycle scripts on packages
that requires it.

Our design principle here is that by separating the install
into these distinctive two steps we can grant the user power
to control what lifecycle scripts they want to run and most
importantly the ability to block unwanted ones.

To improve the UX around this, an install will persist the choice
of allowed or blocked packages to the lockfile so that the user
will only need to make that decision once for a given package in
the same registry and also allowing that decision to be shared
with a team by checking in the lockfile to a version control
system.

How it works:
- Running `vlt install` no longer run lifecycle scripts
- Running `vlt query :scripts` allows the user to inspect what
packages currently installed in their system rely on post install
lifecycle scripts
- Running `vlt build --scope=<query>` allows the user to run
lifecycle scripts to only the matching packages from the given
query and adds these packages to an allowed list in the lockfile
- Running `vlt build skip --scope=<query>` skips the lifecycle
sciprts of the matching packages using the given query and adds
these packages to a block list in the lockfile.
- Running `vlt build` with no parameters will run lifecycle
scripts to all packages on the previous install
- Running `vlt build skip` with no parameters will block lifecycle
scripts of all packages on the previous install

Running `vlt install && vlt build` is the functional equivalent
of the prior `vlt install` behavior.