graph: add edge names to internal hoist (#1251)
We have observed specific dependencies relying on using aliased names
from the internal hoist, this changeset replaces usage of the node name
with the edge names for the internal hoist symlink.

This logic change matchs the `pnpm` implementation that enables
`nitro-nightly@3.0.0-20251010-091540-0e3dbc69` to work in our current
install graph.