graph: fix lockfile optional deps
The current lockfile format is meant to be shared across a team.
That means that it should retain information about optional
dependencies that might not have been installed as part of an
install.

This changeset fixes a regression introduced in
1a6fc52d7b7355a730f0d1cb955944a75614a9b9 that caused optional
dependencies that were not installed as part of a `vlt install`
to be excluded from the final `vlt-lock.json` file.