query: fix remove edges helper
Some selectors like `:peer` were omitting results due to the
`removeEdge` helper function greedly removing the node from
an edge, despite the fact that node could be fulfilling other
edges.

This changeset fixes it by properly limitting the ability of
`removeEdge` to only remove a node in case that specific edge
is the only edge that links into that node and comparing all
edges for usage in the `removeUnlinkedNodes` helper function.