bugfix: `use-dashboard-root-check` fires multiple times
there was multiple flashes of the `/dashboard` view when navigating via
a `dashboard-item` project. This was beacuse `use-dashboard-root-check`
had the `stamp` as a dependency, and when navigating a `dashboard-item`,
it also updated the stamp.

to fix this, just pass `any` value as a dependency to `use-dashboard-root-check`, such as `forceRerender`,
which will be used as an invalidator.