gui: bugfixes and loading states! (#1293)
## Overview

This PR introduces a loading state on the `explorer` when navigating /
clicking a `dashboard-item` on the `/dashboard` view -> to the
transition of the `selected-item` on the `/explore` view

As part of the feature of the loading state, it fixes a lot of rendering
/ reloading bugs that came up:

<details open>
<summary>
Before + After
</summary>

### Before

https://github.com/user-attachments/assets/df45802a-5a29-444d-8766-34ea44c7e21a

### After

https://github.com/user-attachments/assets/e567deaa-f002-450f-b8a7-ef4e3da8b38c
</details>

### `use-dashboard-root-check` fires

This hook check, used for checking the state of the `/dashboard` if
there are configured `dashboard-roots` in the `config`, fired multiple
times, by design initially when the `stamp` changed.
The `stamp` was fired to be updated in the `@/app/settings.tsx`
initially, whenever a `dashboard-root` was updated.

However the `stamp` is also updated when a user clicks on a
`dashboard-item` as part of the `request-route-transition` module.

To fix this, we remove `stamp` as a dependency from this hook, and
instead pass `any` `rerender` invalidator.

This fixes the `/dashboard` FOC when navigating via a `dashboard-item`

### `request-route-transition`

Removes an old `window.scroll` navigation feature that was leftover from
before we had a proper router in the gui.

### `selected-item` has stale content

#### The bug

Sometimes during a navigation from a `dashboard-item`, `selected-item`
on the `/explore` presents with stale data of the previous item with a
FOC.

This was because the graphs `edges` and `nodes` were stored in a global
`zustand` state (useGraphStore). When initially navigating to an item,
there is no stale data (this is the first graph loaded), however another
project navigation had stale data for the first render, because the old
`edges` and `nodes` were still in memory. This temporal coupling caused
component to render with mismatched data between each step.

Additionally, without a loading state or loading guard, components would
render whatever was in the store, even if it was indeed stale, and
components would show `no-results` for empty arrays, even during
loading.

More race conditions were introduced, because navigation happened before
the data was ready: route changed -> component mounted -> then data
fetching

#### The fix

Introduce a new state called `graphStamp` in the `zustand` graph store,
which exists to detect stale data during async loading:

- stamp = "Project-A" (what the user wants to see)
- graph data = "Project-B" (what's currently loaded)

Additionally we assign some `key` to the `selected-item` to ensure that
the `item.id` matches

### `inProgress` state on the `dashboard-grid`

Because we handle the loading state on the `/explore` view, we can
remove this `inProgress` state here.

Closes #1287