graph: improved peer deps support (#1278)
Improves peer dependency support by introducing new logic to better
handle their reusable nature.

A new `src/graph/src/ideal/peers.ts` module that defines and manages
sets of peer contexts has been introduced and is now tightly coupled
with the logic that traverses manifest metadata from
`src/graph/src/append-nodes.ts` to create a new system that better
adheres to peer dependencies.

### Peer Dependencies Support

Improved support to peer dependencies works by resolving peer
dependencies to known versions of that dependency already in used by
ancestor nodes in the graph, the algorithm keeps track of compatible
versions by building sets of known dependencies, when version range
declarations conflict and no intersecting version can be found then new
set is going to be created and used to resolve dependencies from that
node and its children.

When multiple sets are defined due to incompatible version found, it's
possible that intermediate nodes get duplicated, here is a very
simplistic example to illustrate the idea:

```mermaid
flowchart TD
  
  subgraph c has peer deps
    R("root") --> A
    R --> B
    A("a") --> C1("c")
    A --> P1
    B("b") --> C2("c")
    B --> P2
    C1 --> P1("p@1")
    C2 --> P2("p@2")
    
  end
  
  subgraph no peer deps
    r("root") --> a
    r --> b
    a --> p1("p@1")
    a --> c
    b --> c
    b --> p2("p@2")
  end
```

In the example above the dependency `c` is resolved to an unique node
when it has no peer dependencies but when a dependency on `p@1 || 2` is
added then it's then duplicated in order to reuse the dependency sets of
both `a` and `b`.

Fixes: https://github.com/vltpkg/vltpkg/issues/86
Fixes: https://github.com/vltpkg/vltpkg/issues/873