graph: add linked folders management (#1321)
This commit introduces the ability to manage dependencies of linked
folder dependencies within a given project.

Similar to the root project and workspaces, changes to the
`package.json` file of linked folders will be reflected on the next `vlt
install`. It's also possible to add and remove dependencies to nested
folders of a project by navigating to that directory and running `vlt
install <pkg>` (or uninstall to remove).

Fixes: https://github.com/vltpkg/vltpkg/issues/781
Closes: https://github.com/vltpkg/vltpkg/pull/815