dep-id: always include registry names
Always include the registry name in registry dep-ids.

This change opens the way to more easily replace the default
registry in the future, along with cleaning up a bunch of
edge case scenarios when handling different variations of
custom vs default registry and fixing a bug where packages
would be duplicated in the `node_modules/.vlt` store folder
due to duplicated dep-ids to represent the same package.

Fixes: https://github.com/vltpkg/vltpkg/issues/1328