package-info: add remote integrity checks
Add the ability for remote type of dependencies (these are usually
random tarball url definitions or git resolutions with known tarball
artifacts to download from) to store an integrity value calculated
the first time they're download and the ability to check if that
integrity value matches anytime you're installing from a lockfile
that holds that integrity value.

This adds an extra layer of checks when installing a project from
a lockfile with a known remote resolution.