gui: perf improvements for readme fetch (#1345)
## Overview

Because we currently fetch the `readme` from an external source if its
not a local package, the time to make that request can take
significantly longer than the time to fetch the manifests and
packuments, and process the details for the `selected-item`. As a result
we're shown an extra long loading state on the `selected-item`, just for
waiting on `readme`.

To fix this, we can extract out the fetch `readme` logic into a seperate
module and introduce a loading state for the readme itself:
`isLoadingReadme`, and show a root loading state as well as a loading
state for the `tabs-overview` (where the readme is rendered).

### Other fixes

Beacuse we introduced a new state for the `isLoadingReadme` it arose a
problem with updated the mock store states, so a fixture was created for
general 'empty' states, and all tests using the `useSelectedItemStore`
had the mock state updated for this fixture to keep things dry.