[bugfix] gui: contributor query navigation with special chars (#1352)
## Overview

This PR fixes query navigation when clicking a contributor in the GUI if
the contributor email has special chars `+`, and addresses a case when
if the contributor has a ` ` in the name. Additionally, it prioritizes
querying through emails over names if both are present. It extracts out
the `handleQueryContributor` handler from the `Contributor` component
and writes simple tests to ensure the behaviour remains consistent.